+++ CAL_PROJETO/src/Node.h
@@ -8,18 +8,18 @@ using namespace std;
 
 class Node {
 protected:
-	long long id;
+	int id;
 	Localizacao coords;
 
 public:
 	Node();
 	~Node();
-	Node(long long id, double lon, double lat);
-	long long getID() const;
+	Node(int id, double lon, double lat);
+	int getID() const;
 	double getLongitude() const;
 	double getLatitude() const;
 	double distanceTo(const Node & n2) const;
-	void setID(long long id);
+	void setID(int id);
 	void setLocation(Localizacao loc);
 	void operator=(const Node & n2);
 	bool operator==(const Node & n2) const;
@@ -36,7 +36,7 @@ inline ostream& operator <<(ostream & o, Node n)
 inline istream& operator >>(istream & i, Node & n) {
 	char b1{}, b2{};
 	Localizacao loc{};
-	long long id{};
+	int id{};
 
 	i >> id >> b1 >> loc >> b2;
